% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ER injury explorer Shiny app}
\usage{
run_app()
}
\value{
No return value. This function is called for its side effect of
starting a Shiny app.
}
\description{
Launches the Shiny application bundled with this package.
}
